Refactor archive/egia code, add tracy, add download tab

+++ src/storage/game/ArchiveVersion.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 
 namespace EGL3::Storage::Game {
-    enum class ArchiveVersion : uint16_t {
+    enum class ArchiveVersion : uint8_t {
         Initial = 0,
 
         LatestPlusOne,
